+++ squidlib-util/src/main/java/squidpony/squidai/astar/Heuristic.java
@@ -41,5 +41,5 @@ public interface Heuristic<N> {
 	 * @param node the start node
 	 * @param endNode the end node
 	 * @return the estimated cost */
-	float estimate(N node, N endNode);
+	double estimate(N node, N endNode);
 }
